refactored fucntion to make it more readable
