+++ db/queries/assessment_records/_helpers.py
@@ -1,3 +1,4 @@
+from collections import defaultdict
 from datetime import datetime
 
 import jsonpath_rw_ext
@@ -5,6 +6,8 @@ import requests
 from config.mappings.assessment_mapping_fund_round import (
     fund_round_data_key_mappings,
 )
+from db.models.assessment_record import TagAssociation
+from flask import current_app
 
 
 def get_answer_value(application_json, answer_key):
@@ -228,3 +231,54 @@ def update_tag_associations(assessment_metadatas):
                 tag_associations
             )
     return assessment_metadatas
+
+
+def get_existing_tags(application_id):
+    """Queries the database for existing tags associated with a specific
+    application, organises them by tag ID, and returns a defaultdict where each
+    tag ID is associated with a list containing tuples of creation timestamps and
+    their corresponding tag instances.
+
+    Example:
+        {'0000000-00000-00000-000001':
+        [(datetime.datetime(2023, 11, 17, 18, 48, 25, tzinfo=datetime.timezone.utc),
+          <TagAssociation db91a66b->), .......
+          ]}
+
+    """
+    existing_tags = TagAssociation.query.filter(
+        TagAssociation.application_id == application_id,
+    ).all()
+    list_existing_tags = defaultdict(list)
+    for existing_tag in existing_tags:
+        tag_id = str(existing_tag.tag_id)
+        list_existing_tags[tag_id].append(
+            (existing_tag.created_at, existing_tag)
+        )
+    return list_existing_tags
+
+
+def filter_tags(incoming_tags, existing_tags):
+    """If an incoming tag ID matches an existing tag ID, it skips that tag;
+    otherwise, it appends the tag list from the existing tags to the filtered tags
+    list."""
+    filtered_tags = []
+    for (
+        existing_tag_id,
+        existing_tag_list,
+    ) in existing_tags.items():
+        _incoming_tag = next(
+            (
+                incoming_tag
+                for incoming_tag in incoming_tags
+                if incoming_tag.get("id") == existing_tag_id
+            ),
+            None,
+        )
+        if _incoming_tag:
+            current_app.logger.info(
+                f"Tag id is already associated: {existing_tag_id}"
+            )
+        else:
+            filtered_tags.append(existing_tag_list)
+    return filtered_tags
